LibWeb: Move canvas `DrawingState` to standalone file
This is required to avoid a circular dependency in a following commit
