+++ Libraries/LibWeb/HTML/Canvas/CanvasShadowStyles.h
@@ -35,8 +35,8 @@ protected:
     CanvasShadowStyles() = default;
 
 private:
-    CanvasState::DrawingState& my_drawing_state() { return static_cast<IncludingClass&>(*this).drawing_state(); }
-    CanvasState::DrawingState const& my_drawing_state() const { return static_cast<IncludingClass const&>(*this).drawing_state(); }
+    DrawingState& my_drawing_state() { return static_cast<IncludingClass&>(*this).drawing_state(); }
+    DrawingState const& my_drawing_state() const { return static_cast<IncludingClass const&>(*this).drawing_state(); }
 };
 
 }
