+++ Libraries/LibWeb/HTML/Canvas/CanvasState.cpp
@@ -47,29 +47,4 @@ bool CanvasState::is_context_lost()
     return m_context_lost;
 }
 
-NonnullRefPtr<Gfx::PaintStyle> CanvasState::FillOrStrokeStyle::to_gfx_paint_style()
-{
-    return m_fill_or_stroke_style.visit(
-        [&](Gfx::Color color) -> NonnullRefPtr<Gfx::PaintStyle> {
-            if (!m_color_paint_style)
-                m_color_paint_style = Gfx::SolidColorPaintStyle::create(color).release_value_but_fixme_should_propagate_errors();
-            return m_color_paint_style.release_nonnull();
-        },
-        [&](auto handle) {
-            return handle->to_gfx_paint_style();
-        });
-}
-
-Gfx::Color CanvasState::FillOrStrokeStyle::to_color_but_fixme_should_accept_any_paint_style() const
-{
-    return as_color().value_or(Gfx::Color::Black);
-}
-
-Optional<Gfx::Color> CanvasState::FillOrStrokeStyle::as_color() const
-{
-    if (auto* color = m_fill_or_stroke_style.get_pointer<Gfx::Color>())
-        return *color;
-    return {};
-}
-
 }
