+++ Libraries/LibWeb/HTML/Canvas/CanvasTextDrawingStyles.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <LibWeb/HTML/Canvas/CanvasState.h>
+#include <LibWeb/HTML/Canvas/DrawingState.h>
 
 namespace Web::HTML {
 
@@ -37,8 +37,8 @@ protected:
     CanvasTextDrawingStyles() = default;
 
 private:
-    CanvasState::DrawingState& my_drawing_state() { return static_cast<IncludingClass&>(*this).drawing_state(); }
-    CanvasState::DrawingState const& my_drawing_state() const { return static_cast<IncludingClass const&>(*this).drawing_state(); }
+    DrawingState& my_drawing_state() { return static_cast<IncludingClass&>(*this).drawing_state(); }
+    DrawingState const& my_drawing_state() const { return static_cast<IncludingClass const&>(*this).drawing_state(); }
 };
 
 }
